// Project/dff_library/ui_components/components/BasicMicroInteractions/utils/enum.ts
export enum BasicMicroInteractionTypes {
  CheckMark = 'Check Mark',
  CheckMarkWhite = ' Check Mark White',
  Delete = 'Delete',
  FavouriteAdd = 'Favourite Add',
  FavouriteRemove = 'Favourite Remove',
  Extra = 'Extra',
  IManLoader = 'IManLoader',
  TransitionaryAnimation = 'Transitionary Animation',
}